% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_semicolon.R
\name{add_semicolon}
\alias{add_semicolon}
\title{Parses Set Score}
\usage{
add_semicolon(xstr)
}
\arguments{
\item{xstr}{a string of point outcomes}
}
\description{
Adds semicolons to unprocessed strings of game outcomes in the \code{point_by_point} data set
}

