% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_players.R
\name{atp_players}
\alias{atp_players}
\title{Biographic Details of ATP Players}
\format{A data frame with 52,167 rows and 9 variables}
\source{
\url{https://github.com/JeffSackmann/tennis_atp}
}
\description{
This dataset contains the names and basic demographic details of current and past players for the ATP World Tour. The variables of the dataset are:
}
\details{
\itemize{
  \item player_id. A numeric id to link players to other datasets
  \item first_name. A character vector of player first names
  \item last_name. A character vector of player last names
  \item player_id. A numeric id to link players to other datasets
  \item hand. A character indicating handedness type: L = left, R = right, U = unknown, A = ambidextrous
  \item birthdate. A numeric birthdate, YYYYYMMDD
  \item country_code. A character of 3-letter codes for country
 \item invalid. Logical if whether player names has any blanks or missing values.
  \item name. A character of player full names, first name + last name
  \item dob. Date of birth as POSIXct class
}
}

