% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_rankings.R
\name{atp_rankings}
\alias{atp_rankings}
\title{Rankings of ATP Players}
\format{A data frame with 2,593,632 rows and 5 variables}
\source{
\url{https://github.com/JeffSackmann/tennis_atp}
}
\description{
This dataset contains the ranking history of current and past players for the ATP World Tour. The variables of the dataset are:
}
\details{
\itemize{
  \item ranking_date. A numeric date of rank, YYYYYMMDD
  \item ranking. A numeric rank
  \item player_id. A numeric id to link players to other datasets
  \item rankings_points. A numeric point value for rank
  \item date. Date of ranking as POSIXct class
}
}

