% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_tournaments2.R
\name{atp_tournaments}
\alias{atp_tournaments}
\title{ATP Tournaments}
\format{A data frame with 67 rows and 8 variables}
\source{
\url{http://www.atpworldtour.com/en/tournaments}
}
\description{
This dataset contains information on the 2016 ATP tournament calendar:
}
\details{
\itemize{
  \item name. A character name of the tournament
  \item location. A character venue of tournament
  \item level. A character name of the tounament level
  \item start_date. A date-time object for the start of the tournament
  \item end_date. A date-time object for the end of the tournament
  \item single_draw. A character vector of the size of the singles draw
  \item surface. A character description of the tournament surface
  \item prize. A character of the total prize money to earn
}
}

