\name{deuce-package}
\alias{deuce-package}
\alias{deuce}
\docType{package}
\title{
Download ATP player and match statistics
}
\description{
This package provides data frames of match-specific and yearly match summary
    statistics for players on ATP tour and functions to download match-, year-
    and surface-specific statistics from the Open Era to the present.
}
\details{
\tabular{ll}{
Package: \tab ATP\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 07-31-2014\cr
License: \tab GPL-2\cr
}
}
\author{

Maintainer: Stephanie Kovalchik <s.a.kovalchik@gmail.com>

}

\keyword{ package }

