% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_prediction.R
\name{elo_prediction}
\alias{elo_prediction}
\title{Elo-Based Win Prediction}
\usage{
elo_prediction(rating1, rating2)
}
\arguments{
\item{rating1}{Numeric of Player 1 Elo rating}

\item{rating2}{Numeric of Player 2 Elo rating}
}
\value{
Match win probability for Player 1
}
\description{
This function returns the prediction for a match based on player Elo ratings
}
\examples{
elo_prediction(2000, 1842)

}
