% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_atp_ranking.R
\name{fetch_atp_rankings}
\alias{fetch_atp_rankings}
\title{Download Rankings from ATP Tour Website}
\usage{
fetch_atp_rankings(date, min_rank = 1, max_rank = 100)
}
\arguments{
\item{date}{A character date for extracting rankings, YYYY-MM-DD}

\item{min_rank}{A numeric for the highest ranked player to include}

\item{max_rank}{A numeric for the lowest ranked player to include}
}
\description{
This function extracts ATP singles rankings for a particular date and ranking group
}

