% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_tournaments.R
\name{fetch_atp_tournaments}
\alias{fetch_atp_tournaments}
\title{Download Current ATP Calendar}
\usage{
fetch_atp_tournaments(challenger = FALSE)
}
\arguments{
\item{challenger}{Logical indicator if fetching challenger or World Tour calendar}
}
\description{
This function downloads the ATP calendar
}

