% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_odds.R
\name{fetch_odds}
\alias{fetch_odds}
\title{Fetch Odds Data for Slam Matches}
\usage{
fetch_odds(atp = T)
}
\arguments{
\item{atp}{Logical; True if Men's ATP match, False if not}
}
\description{
This function extracts official match data from Tennis-data.co.uk
}
\examples{
 atp_odds <- fetch_odds(atp = TRUE)
 wta_odds <- fetch_odds(atp = FALSE)

}
