\name{fetch_player_code}
\alias{fetch_player_code}

\title{
Extract ATP player code for matchfacts
}
\description{
Returns the id for a player as represented in ATP site databases
}

\usage{
fetch_player_code(player, year = NULL)
}

\arguments{
   \item{player}{character name of player}
   \item{year}{numeric year}
}


\value{ Character code for player}


\references{
\url{http://www.atpworldtour.com}
}

\author{
Stephanie Kovalchik <s.a.kovalchik@gmail.com>
}


