% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_repo_data.R
\name{fetch_repo_data}
\alias{fetch_repo_data}
\title{Fetch Repo Data}
\usage{
fetch_repo_data(file, ...)
}
\arguments{
\item{file}{File path of desired csv}

\item{...}{other options passed to fetch_repo_data}
}
\value{
A csv file read in as desired
}
\description{
This function fetches a csv from the desired GitHub repo to be used in match and player analysis
}
