% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_repo_files.R
\name{fetch_repo_files}
\alias{fetch_repo_files}
\title{Scrapes CSV files from GITHUB repository}
\usage{
fetch_repo_files(url)
}
\arguments{
\item{.}{url The repo URL address}
}
\description{
Scrapes CSV files from GITHUB repository
}
