% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_slam_pbp.R
\name{fetch_slam_pbp}
\alias{fetch_slam_pbp}
\title{Download IBM Slamtracker Point-by-Point Data}
\usage{
fetch_slam_pbp(event, year)
}
\arguments{
\item{event}{Character value of slam and one of: 'ausopen', 'wimbledon', 'usopen', or 'frenchopen'}

\item{year}{Numeric year}
}
\value{
data frame of point by point results
\itemize{
 \item PointServer. Numeric indicator of player who is serving
 \item PointWinner. Numeric indicator of player who won the point
 \item Set. Numeric of set number
 \item Game. Numeric of game number
 \item Time. Elapsed time at start of point (hh:mm:ss)
 \item Ace. Numeric indicator if server made and ace
 \item Winner. Numeric indicator if point ended with a winner
 \item DoubleFault. Numeric indicator if point ended in double fault
 \item UnforcedError. Numeric indicator if point ended in an unforced error
 \item NetPoint. Numeric indicator if point was a net point
 \item WinnerShotType. Character of type of shot for winner (F or B)
 \item ServerError. Logical if server made an error
 \item ReceiverError. Logical if receiver made an error 
 \item SpeedMPH. Numeric of MPH of serve
 \item ServeNumber. Numeric indicator of first or second serve
 \item BreakPointWon. Numeric indicator if break point was won
 \item BreakPointOpportunity. Numeric indicator if had a break point opportunity
 \item GameWinner. Numeric if point ended in game winner and for which player
 \item ForcedError. Numeric of whether point ended in forced error
 \item MatchCode. Identifier of the match
 \item Round. Numeric value of round (1 = First, 7 = Final)
 \item ServerName. Character of server's name
 \item Player1. Character of Player 1 name
 \item Player2. Character of PLayer 2 name
}
}
\description{
This function extracts historical point-by-point data for Grand Slam events
}
\details{
This has been tested for 2017. Previous years may not be available as data is archived or site domains have changed.
}
