% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_tour_stats.R
\name{fetch_tour_stats}
\alias{fetch_tour_stats}
\title{Download Performance Statistics for ATP}
\usage{
fetch_tour_stats(stat = "Aces", year = "all", surface = "Clay")
}
\arguments{
\item{stat}{A character variable that is one of the following: "Aces", "1st Serve", "1st Serve Won", "2nd Serve #' Won", "Service Games Won",  "Break Points Saved",  "1st Serve Return Won", "2nd Serve Return Won", or "Break Points Converted")}

\item{year}{A character for the year or "all" for career stats}

\item{surface}{A character for the surface type (Clay, Hard, or Grass); use "all" for all surfaces}
}
\description{
This function extracts performance statistics for the ATP
}

