% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iid_player_probs.R
\name{iid_player_probs}
\alias{iid_player_probs}
\title{Generate IID Conditional Probability Inputs}
\usage{
iid_player_probs(pa, pb, advantage)
}
\arguments{
\item{pa}{The probability of winning on serve for player A}

\item{pb}{The probability of winning on serve for player B}

\item{advantage}{A logical value indicating whether the match is decided with an advantage set}
}
\description{
This function creates the probabilities for game, set, and match wins for the serving player given their probability of winning a point on serve
}

