% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Calculate Importance of Point}
\usage{
importance(point_a, point_b, game_a, game_b, set_a, set_b, serving_player = 1,
  matrices, bestof3 = TRUE, advantage = TRUE)
}
\arguments{
\item{point_a}{The current point score within a game for player A (numerical; 1 = 15, 2 = 30, etc.)}

\item{point_b}{The current point score within a game for player B}

\item{game_a}{The current game score within a game for player A}

\item{game_b}{The current game score within a game for player B}

\item{set_a}{The current set score within a game for player A}

\item{set_b}{The current set score within a game for player B}

\item{serving_player}{Numerical (1 or 2) if first element in \code{matrices} is the serving player = 1; 2 otherwise}

\item{matrices}{List of IID inputs from \code{iid_player_probs}}

\item{bestof3}{Logical indicator if a best-of-3 match}

\item{advantage}{Logical indicator if an advantage match}
}
\description{
Calculates the importance of a point using Carl Morris' definition of importance. This is the difference in the probability of winning a match when the current point is won compared to when it is lost.
}

