% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_win.R
\name{in_match_win}
\alias{in_match_win}
\title{Estimating Conditional Match Win}
\usage{
in_match_win(point_a, point_b, game_a, game_b, set_a, set_b,
  serving_player = 1, matrices, bestof3 = TRUE, advantage = TRUE)
}
\arguments{
\item{point_a}{The current point score within a game for player A (numerical; 1 = 15, 2 = 30, etc.)}

\item{point_b}{The current point score within a game for player B}

\item{game_a}{The current game score within a game for player A}

\item{game_b}{The current game score within a game for player B}

\item{set_a}{The current set score within a game for player A}

\item{set_b}{The current set score within a game for player B}

\item{serving_player}{Numerical (1 or 2) if first element in \code{matrices} is the serving player = 1; 2 otherwise}

\item{matrices}{List of IID inputs from \code{iid_player_probs}}

\item{bestof3}{Logical indicator if a best-of-3 match}

\item{advantage}{Logical indicator if an advantage match}
}
\description{
This function uses IID formula to estimate the conditional probability that the player who is currently serving wins the match given the match format and current score
}

