% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{pbp}
\alias{pbp}
\title{Create Point-by-point Data}
\usage{
pbp(obj)
}
\arguments{
\item{obj}{a row of the \code{point_by_point} data frame}
}
\description{
Creates an expanded point-level version from the flat representation in \code{point_by_point}
}

