% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_loss_return.R
\name{return_loss_importance}
\alias{return_loss_importance}
\title{Calculate Loss Importance for a Point}
\usage{
return_loss_importance(id, serve.prob, return.prob, bestof3 = TRUE,
  advantage = TRUE)
}
\arguments{
\item{id}{Score id for the start of the current point}

\item{return.prob}{Numeric serve win prob of current returner}

\item{bestof3}{Logical indicator if best-of-3 match (TRUE) or best-of-5 (FALSE)}

\item{advantage}{Logical if advantage set match (TRUE) or tiebreak match (FALSE)}

\item{server.prob}{Numeric serve win prob of current server}
}
\value{
Numeric value of importance
}
\description{
This function calculates the match win prediction for matches in progress based on the current score
}
