% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_match_data.R
\name{tidy_match_data}
\alias{tidy_match_data}
\title{Tidy Match Data}
\usage{
tidy_match_data(data, atp = T)
}
\arguments{
\item{data}{A data table containing Match data}

\item{atp}{Logical, TRUE if ATP match, FALSE if WTA match}
}
\description{
This function tidy's match data at the game level
}
\examples{
\dontrun{
wta_matches_tidy <- tidy_match_data(wta_matches, atp = F)
} 
}
