% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_player_data.R
\name{tidy_player_data}
\alias{tidy_player_data}
\title{Tidy Player Data}
\usage{
tidy_player_data(data)
}
\arguments{
\item{data}{A data table containing player data}
}
\description{
This function tidy's the player's name to by uniform across all imported data
}
