% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_slam_point_by_point.R
\name{tidy_slam_point_by_point}
\alias{tidy_slam_point_by_point}
\title{Tidy Slam Point by Point}
\usage{
tidy_slam_point_by_point(points, matches)
}
\arguments{
\item{points}{A data set with detail of every point in a match}

\item{matches}{A data set with detail of every match}
}
\description{
This function takes two data sets and returns a combined data set with every point in every match
}
\examples{
\dontrun{
tidy_slam_point_by_point(point_by_point, point_by_matches)
}
}
