% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.R
\name{update_elo}
\alias{update_elo}
\title{Update Elo Ratings}
\usage{
update_elo(elo, match, win, slam = TRUE)
}
\arguments{
\item{elo}{A two-element vector of player Elo ratings at the start of a match}

\item{match}{A two-element vector of the career matches played by each player}

\item{win}{A two-element vector with a 1 for the winner and 0 for the loser}

\item{slam}{A logical indicator of whether the match was a Grand Slam}
}
\description{
This function updates player Elo ratings given the outcome of a match
}

