% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_score.R
\name{update_score}
\alias{update_score}
\title{Update Score}
\usage{
update_score(pointa, pointb, gamea, gameb, seta, setb, advantage, bestof3)
}
\arguments{
\item{pointa}{Points won by server}

\item{pointb}{Points won by receiver}

\item{gamea}{Games won by server}

\item{gameb}{Games won by receiver}

\item{seta}{Sets won by server}

\item{setb}{sets won by receiver}

\item{advantage}{Indicator (T or F) if advantage final set}

\item{bestof3}{Indicator (T or F) if best of 3 match or best of 5 match}
}
\value{
Match win probability for serving player
}
\description{
This function updates the score if game, set, or match won
}
