% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atp_tournaments2.R
\name{atp_tournaments}
\alias{atp_tournaments}
\title{ATP Tournaments}
\format{A data frame with 127 rows and 9 variables}
\source{
\url{http://www.atpworldtour.com/en/tournaments}
}
\description{
This dataset contains events for the 2017 ATP World Tour and ATP Challenger Tour:
}
\details{
\itemize{
  \item name. A character name of the tournament
  \item location. A character venue of tournament
  \item level. A character name of the tounament level
  \item start_date. A date-time object for the start of the tournament
  \item end_date. A date-time object for the end of the tournament
 \item draw. A numeric vector of the size of the singles draw
 \item matches. A numeric vector of the number of planned matches 
  \item surface. A character description of the tournament surface
 \item tier. A character cateogry for the tournament series
  \item prize. A character of the total prize money to earn
}
}

