% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_wta_draws.R
\name{fetch_draw}
\alias{fetch_draw}
\title{Download Grand Slam Draws}
\usage{
fetch_draw(tournament, year, atp = TRUE)
}
\arguments{
\item{tournament}{Name of Grand Slam}

\item{year}{Numeric year of event}

\item{atp}{Logical indicator of ATP or WTA}
}
\value{
data frame of draw results
\itemize{
 \item{"winner"}{Name of player who won}                            
 \item{"loser"}{Name of player who lost}                      
 \item{"round"}{Numeric round (7 = Final, 1 = First Round)}                      
 \item{"date"}{Date of start of tournament}            
 \item{"year"}{Numeric of year}              
 \item{"tournament"}{Name of tournament}                
}
}
\description{
This function downloads Grand Slam tournament draws for ATP and WTA singles
}
\examples{
fetch_match_stats("Australian Open", 2017, atp = FALSE)

}

