% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_head_to_head.R
\name{fetch_head_to_head}
\alias{fetch_head_to_head}
\title{Download Head-to-Head}
\usage{
fetch_head_to_head(player1, player2, atp = TRUE)
}
\arguments{
\item{player1}{Character name of player (must match name used by source)}

\item{player2}{Character name of player (must match name used by source)}

\item{atp}{Logical indicator for ATP or WTA}
}
\value{
data frame of match activity and results
\itemize{
 \item year. Numeric of year of match
 \item tournament. Character name of tournament
 \item round. Character of round of match
 \item surface. Character of surface
 \item winner. Character name of winner of match
 \item loser. Character name of loser of match
 \item score. Character of score
}
}
\description{
This function extracts head-to-heads for WTA and ATP players from \url{http://www.stevegtennis.com}
}
\examples{
fetch_head_to_head("Rafael Nadal", "Roger Federer")

}
