% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_match_stats.R
\name{fetch_match_stats}
\alias{fetch_match_stats}
\title{Download Match Statistics}
\usage{
fetch_match_stats(player, year)
}
\arguments{
\item{player}{Character name of ATP player}

\item{year}{Numeric year or "all" for all years}
}
\value{
data frame of match activity and results
\itemize{  
 \item Aces.1. Count of aces for player 1                             
 \item AcesPercentage.1. Percentage of aces out of total aces in match                       
 \item DoubleFaults.1. Count of double faults for player 1                       
 \item DoubleFaultsPercentage.1. Percentage of double faults in match             
 \item FirstServePercentage.1. First serve percentage for player 1               
 \item FirstServeDividend.1. Numerator for first serve percentage for player 1                 
 \item FirstServeDivisor.1. Denominator for first serve percentage for player 1                  
 \item FirstServePointsWonPercentage.1. Percentage of first serve points won for player 1      
 \item FirstServePointsWonDividend.1. Numerator for first serve points won for player 1        
 \item FirstServePointsWonDivisor.1. Denominator for first serve points won for player 1         
 \item SecondServePointsWonPercentage.1. Percentage of second serve points won for player 1     
 \item SecondServePointsWonDividend.1. Numerator for second serve points won for player 1       
 \item SecondServePointsWonDivisor.1. Denominator for second serve points won for player 1        
 \item BreakPointsSavedPercentage.1. Break points saved percentage for player 1         
 \item BreakPointsSavedDividend.1. Numerator for break points saved percentage           
 \item BreakPointsSavedDivisor.1. Denominator for break points saved percentage            
 \item ServiceGamesPlayed.1. Count of service games played for player 1                 
 \item ServiceGamesPlayedPercentage.1. Percentage of service games played out of match       
 \item FirstServeReturnPointsPercentage.1. Percentage of first serve return for player 1   
 \item FirstServeReturnPointsDividend.1. Numerator of first serve return percentage     
 \item FirstServeReturnPointsDivisor.1. Denominator of first serve return percentage      
 \item SecondServePointsPercentage.1. Percentage of second serve for player 1        
 \item SecondServePointsDividend.1. Numerator of second serve percentage          
 \item SecondServePointsDivisor.1. Denominator of second serve percentage           
 \item BreakPointsConvertedPercentage.1. Percentage of break points converted for player 1     
 \item BreakPointsConvertedDividend.1. Numerator of break points converted       
 \item BreakPointsConvertedDivisor.1. Denominator of break points converted        
 \item ReturnGamesPlayed.1. Total of return games played for player 1                  
 \item ReturnGamesPlayedPercentage.1. Percentage of return games played out of match        
 \item TotalServicePointsWonPercentage.1. Percentage of total service points won for player 1    
 \item TotalServicePointsWonDividend.1. Numerator for total service points won for player 1      
 \item TotalServicePointsWonDivisor.1. Denominator for total service points won for player 1       
 \item TotalReturnPointsWonPercentage.1. Percentage of total return points won for player 1     
 \item TotalReturnPointsWonDividend.1. Numerator of return points won       
 \item TotalReturnPointsWonDivisor.1. Denominator of return points won        
 \item TotalPointsWonPercentage.1. Percentage of total points won by player 1           
 \item TotalPointsWonDividend.1. Numerator for total points won             
 \item TotalPointsWonDivisor.1. Denominator for total points won              
 \item Aces.2. Count of aces for player 2                             
 \item AcesPercentage.2. Percentage of aces out of total aces in match                       
 \item DoubleFaults.2. Count of double faults for player 2                       
 \item DoubleFaultsPercentage.2. Percentage of double faults in match             
 \item FirstServePercentage.2. First serve percentage for player 2               
 \item FirstServeDividend.2. Numerator for first serve percentage for player 2                 
 \item FirstServeDivisor.2. Denominator for first serve percentage for player 2                  
 \item FirstServePointsWonPercentage.2. Percentage of first serve points won for player 2      
 \item FirstServePointsWonDividend.2. Numerator for first serve points won for player 2        
 \item FirstServePointsWonDivisor.2. Denominator for first serve points won for player 2         
 \item SecondServePointsWonPercentage.2. Percentage of second serve points won for player 2     
 \item SecondServePointsWonDividend.2. Numerator for second serve points won for player 2       
 \item SecondServePointsWonDivisor.2. Denominator for second serve points won for player 2        
 \item BreakPointsSavedPercentage.2. Break points saved percentage for player 2         
 \item BreakPointsSavedDividend.2. Numerator for break points saved percentage           
 \item BreakPointsSavedDivisor.2. Denominator for break points saved percentage            
 \item ServiceGamesPlayed.2. Count of service games played for player 2                 
 \item ServiceGamesPlayedPercentage.2. Percentage of service games played out of match       
 \item FirstServeReturnPointsPercentage.2. Percentage of first serve return for player 2   
 \item FirstServeReturnPointsDividend.2. Numerator of first serve return percentage     
 \item FirstServeReturnPointsDivisor.2. Denominator of first serve return percentage      
 \item SecondServePointsPercentage.2. Percentage of second serve for player 2        
 \item SecondServePointsDividend.2. Numerator of second serve percentage          
 \item SecondServePointsDivisor.2. Denominator of second serve percentage           
 \item BreakPointsConvertedPercentage.2. Percentage of break points converted for player 2     
 \item BreakPointsConvertedDividend.2. Numerator of break points converted       
 \item BreakPointsConvertedDivisor.2. Denominator of break points converted        
 \item ReturnGamesPlayed.2. Total of return games played for player 2                  
 \item ReturnGamesPlayedPercentage.2. Percentage of return games played out of match        
 \item TotalServicePointsWonPercentage.2. Percentage of total service points won for player 2    
 \item TotalServicePointsWonDividend.2. Numerator for total service points won for player 2      
 \item TotalServicePointsWonDivisor.2. Denominator for total service points won for player 2       
 \item TotalReturnPointsWonPercentage.2. Percentage of total return points won for player 2     
 \item TotalReturnPointsWonDividend.2. Numerator of return points won       
 \item TotalReturnPointsWonDivisor.2. Denominator of return points won        
 \item TotalPointsWonPercentage.2. Percentage of total points won by player 2           
 \item TotalPointsWonDividend.2. Numerator for total points won             
 \item TotalPointsWonDivisor.2. Denominator for total points won   
 \item Player1. Character name of player 1                              
 \item Player2. Character name of player 2                              
 \item Time. Character match time HH:MM:SS                                 
 \item Tournament. Character name of tournament                           
 \item Round. Character name of round                                
 \item Year. Numeric of year  
}
}
\description{
This function extracts statistics for each player for all ATP, Grand Slam and Olympic events
}
\examples{
\dontrun{
fetch_match_stats("Alexander Zverev", 2017)
}

}
