% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_wta_ranking.R
\name{fetch_wta_rankings}
\alias{fetch_wta_rankings}
\title{Download Current WTA Rankings}
\usage{
fetch_wta_rankings(min_rank = 1, max_rank = 100)
}
\arguments{
\item{min_rank}{A numeric for the highest ranked player to include}

\item{max_rank}{A numeric for the lowest ranked player to include}
}
\value{
data frame of rankings for the specified week
\itemize{
 \item player. Character name of player
 \item nation. Character name of player
 \item points. Numeric of ranking points
 \item tournaments. Numeric of tournaments played
 \item rank. Numeric rank
 \item date. Date object of Monday rankings were updated
}
}
\description{
This function extracts current WTA rankings
}
\examples{
fetch_wta_rankings(1, 250)

}
