% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_match_win.R
\name{in_match_win}
\alias{in_match_win}
\title{Conditional Match Win Prediction}
\usage{
in_match_win(point_a, point_b, game_a, game_b, set_a, set_b, server.prob,
  returner.prob, bestof3 = TRUE, advantage = TRUE)
}
\arguments{
\item{point_a}{Numeric game points won by current server at start of point}

\item{point_b}{Numeric game points won by current returner at start of point}

\item{game_a}{Numeric games won by current server in the current set}

\item{game_b}{Numeric games won by current returner in the current set}

\item{set_a}{Numeric sets won by current server}

\item{set_b}{Numeric sets won by current returner}

\item{server.prob}{Numeric serve win prob of current server}

\item{returner.prob}{Numeric serve win prob of current returner}

\item{bestof3}{Logical indicator if best-of-3 match (TRUE) or best-of-5 (FALSE)}

\item{advantage}{Logical if advantage set match (TRUE) or tiebreak match (FALSE)}
}
\value{
Numeric probability that current server will win the match
}
\description{
This function calculates the conditional match win prediction for matches in progress conditioned on the current score
}
\examples{
in_match_win(3, 2, 5, 5, 0, 0, 0.65, 0.63, bestof3 = T, advantage = F)
in_match_win(2, 3, 5, 5, 0, 1, 0.65, 0.63, bestof3 = T, advantage = F)

}
\references{
Barnett, T. (2013). Teaching probability theory through tennis. Strategic Games.
}
