% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_win.R
\name{match_win}
\alias{match_win}
\title{Match Win Probability}
\usage{
match_win(serve, return, bestof3 = TRUE)
}
\arguments{
\item{serve}{probability that favored player wins a point on serve}

\item{return}{probability that favored player wins a point on return}

\item{bestof3}{logical indicator whether best-of-3 match (TRUE) or best-of-5 (FALSE)}
}
\description{
Match win probability under an IID model
}

