% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wta_importance.R
\name{wta_importance}
\alias{wta_importance}
\title{WTA Point Importance}
\format{A data frame with 2,628 rows and 15 variables}
\description{
This dataset the importance of points for different scorelines in Grand Slam matches:
}
\details{
\itemize{
  \item serve_point. Numeric points won by server in game
  \item return_point. Numeric points won by returner in game
  \item serve_game. Numeric games won by server in set
  \item return_game. Numeric games won by returner in set
  \item serve_set. Numeric sets won by server in match
  \item return_set. Numeric sets won by returner in match
  \item serve_score_name. Nominal point score for server
  \item return_score_name. Nominal point score for returner
  \item game_score. Nominal game score
  \item set_score. Nominal set score
  \item win. Numeric probability of match win for average player serving if point is won
  \item loss. Numeric probability of match win for average player serving if point is  lost
  \item importance. Numeric importance of point
  \item id. Character score id
}
}
